python2/requirement.txt
# Django
django == 1.9
djangorestframework >= 3.3.0
djangorestframework-jwt >= 1.7.2
django-extensions >= 1.5.6
django-import-export >= 0.2.8
django-tinymce >= 2.0.0

# MySQL
MySQL_python == 1.2.5

# xlwt
xlwt == 1.2.0

# xlutils
xlutils == 2.0.0

# openpyxl
openpyxl == 2.4.1

# Support libs
bosonnlp >= 0.7.0
chardet >= 2.3.0
beautifulsoup4 >= 4.3.2
BeautifulSoup >= 2.6.3
jieba >= 0.37
lxml >= 4.0.0
python-dateutil >= 2.5.0
pytz >= 2015.7
requests >= 2.2.1
selenium >= 2.44.0
xlrd >= 0.9.4
docx >= 0.2.4
Pillow >= 4.3.0
olefile >= 0.44
